project setup and migrated to js
