--- task-list/src/App.tsx
@@ -1,13 +1,0 @@
-
-
-import "./App.css"
-import Background from "./components/background/background.components";
-const App = () => {
-  return (
-    <div className="app-container">
-      <Background />
-    </div>
-  );
-}
-
-export default App;
