--- task-list/src/components/background/background.components.tsx
@@ -1,12 +1,0 @@
-import "./background.styles.css"
-import StartTyping from "../start-typing/start-typing.component"
-
-const Background = () => {
-    return (
-        <div className="background-container">
-            <StartTyping />
-        </div>
-    )
-}
-
-export default Background
