--- task-list/src/components/background/background.styles.css
@@ -1,12 +1,0 @@
-.background-container {
-    width: 80%;
-    height: 30vh;
-    border-radius: 2%;
-    margin: 20px;
-    background-image:linear-gradient(to right ,  rgb(178, 12, 255 , 0.85), rgba(2, 187, 255, 0.85)), url('../../../public/peakpx.jpg');
-    background-size: cover;
-    background-position: center;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
