--- task-list/src/components/start-typing/start-typing.component.tsx
@@ -1,15 +1,0 @@
-import "./start-typing.styles.css"
-
-const StartTyping = () => {
-    return(
-        <div className="start-typing-container">
-            <h1>TODO</h1>
-            <div>
-                <input type="text" placeholder="Start typing ..."/>
-                <button>Create</button>
-            </div>
-        </div>
-    )
-}
-
-export default StartTyping
