--- task-list/src/components/start-typing/start-typing.styles.css
@@ -1,33 +1,0 @@
-.start-typing-container{
-    width: 80%;
-}
-.start-typing-container > h1 {
-    letter-spacing: 5px;
-}
-.start-typing-container > div {
-    width: 90%;
-    display: flex;
-    align-items: center;
-    background-color: white;
-    border-radius: 10px;
-    height: 50px;
-    padding: 0px 10px;
-}
-.start-typing-container > div >input {
-    width: 90%;
-    height: 40px;
-    font-size: 15px;
-    outline: none;
-    border: none;
-    border-radius: 10px;
-}
-.start-typing-container > div > button{
-    height: 40px;
-    outline: none;
-    background-color:  rgb(178, 12, 255);
-    color: white;
-    border: none;
-    cursor: pointer;
-    border-radius: 10px;
-    width: 70px;
-}
